Move `OsuHitObjectPositionModifier` to `Utils/`
